image-analysis-app/backend/requirements.txt
fastapi
uvicorn[standard]
pillow
python-multipart
jina-clip
ultralytics  # YOLOv8
pinecone-client
pydantic
python-dotenv
requests
